% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model2}
\alias{model2}
\title{Data from model 2 as proposed in Mukherjee and Zhu(2011)}
\format{A List with 3 elements, a Response Matrix (Y), a Design Matrix (DM), and a true beta matrix (BETA)
\describe{
  \item{Y}{100 X 20 Response Matrix}
  \item{DM}{100 X 50 Design Matrix}
  \item{BETA}{50 X 20 True Beta Matrix}
}}
\source{
\url{https://github.com/zkeller89/RRRR615/blob/master/R/sample_data_edits.R}
}
\usage{
data(model2)
}
\description{
A list containing elements of sample data as specified in Mukherjee and Zhu(2011) as "Model 2", with all singular values of our coefficient matrix equal to 1.
}
\references{
\url{http://dept.stat.lsa.umich.edu/~jizhu/pubs/Mukherjee-SADM11.pdf}
}
\keyword{datasets}

