% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRLM.R
\name{RRLM}
\alias{RRLM}
\title{Create a Reduced Rank Ridge Regression Model}
\usage{
RRLM(X, Y, lambda, r = NULL, CV = NULL, INT = TRUE, Yfit = FALSE)
}
\arguments{
\item{X}{The design matrix for our model}

\item{Y}{The response matrix for our model}

\item{lambda}{The penalizing term for our model}

\item{r}{The rank to which we believe our Beta matrix to be; only r or CV should be specified}

\item{CV}{The proportion of the sum of the singular values needed for our reduced rank; only r or CV should be specified}

\item{INT}{A boolean indicating if an intercept term is desired in our model}

\item{Yfit}{A boolean indicating if fitted Y values and their errors should be returned}
}
\description{
Create a Reduced Rank Ridge Regression Model
}

