\name{RRRR615-package}
\alias{RRRR615-package}
\alias{RRRR615}
\docType{package}
\title{
\packageTitle{RRRR615}
}
\description{
\packageDescription{RRRR615}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RRRR615}
\packageIndices{RRRR615}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RRRR615}

Maintainer: \packageMaintainer{RRRR615}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
